Exploding Zombies!
v16
So ive added some placeholder Animations for the Zombies deaths.
NOW they EXPLODE! hahahahaha :D

Main;
- Added comments for the jars required to start working with NiftyGUI to
create a HUD.
- Changed so that the Bullet no longer fires from the top left corner,
but from the center, or approx of the Sprite.
(This will obviously be changed eventuly to wherever the X/Y Cord of the
gun is?).

assets;
- Bloodsplat is the Placeholder damage indicator for when the Zombie
gets shot.
- zomb#_Death# where # is the number for the zombie or frame, is where
the Placeholder frames for the Death animation are.

HUD;
- Empty Framework Class, and Implementation into Main.

Entity;
- Animation imports for the death and bloodsplats.
- fixed the play on start, and not stoping after animation ends, bugs.
- Realised i can cut down alot of this image stuff by simply assigning
the animation of UP to the other directions then rotating it, instead of
re-creating each and every frame for each animation, least im hoping it
will be an improvement? ^.^

Awsome! so we have the zombies exploding now!
I want to start working on the having the multiple characters for the
players squad/group now, but before that ill need to finaly Merge
accross the Entity of the player into the Entity class since its still
in main (yeh what!?).

Anyway, i keep syaing ill do that, so i better actuly do it.
I think then i might just extend both NPC & PLAYER from ENTITY?

Peace!
- V.