v27 Bullets and Clips!
Added ammunition in Clips now, Just need to add the final weapons to use
the clips and bullets.

Minor edits to the following, in Regards to he Clip Implementation;
- BulletFactory.
- Entity.
- HUD.
- Main.

HUD;
- Now shows realtime information on Clips & Ammo Counts.

Well, got that niffty little bit out of the way, Did not want to play
nice for a while, but it looks like its good to go now.

When players reload, they use up a clip from the available clips, if the
current clip still has ammo, its replaced back into the clips List to be
used later.

- V.